pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dbal"
version = "0.1.0"
description = "Database Abstraction Layer for Data Engineering"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "black>=23.0",
]

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["dbal*", "ingestion*", "fx*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
